move file decrypter to file
